Setting up bridge pattern for class Pusher: - bridge to PusherType (ex: Boris pusher, Vay pusher ...)
A concrete PusherType class has to define:
- move1D, move2D and move3D

Simple inheritance for
- class Pusher1D : public Pusher

A Pusher object is built by the PusherFactory:
- static std::unique_ptr<Pusher>
    createPusher( const GridLayout & layout, const std::string & pusherType )

+++ src/grid/gridlayout.h
@@ -94,9 +94,9 @@ public:
     double ody()const { return dy_ == 0. ? throw error(errorInverseMesh +" dy() (dy==0)"): ody_;}
     double odz()const { return dz_ == 0. ? throw error(errorInverseMesh +" dz() (dz==0)"): odz_;}
 
-    double nbrCellx() const {return nbrCellx_;}
-    double nbrCelly() const {return nbrCelly_;}
-    double nbrCellz() const {return nbrCellz_;}
+    uint32 nbrCellx() const {return nbrCellx_;}
+    uint32 nbrCelly() const {return nbrCelly_;}
+    uint32 nbrCellz() const {return nbrCellz_;}
 
     uint32 nbDimensions() const { return nbDims_ ; }
